Mod: make simple more easy
