+++ examples/helloworld/dubbo/go-client/profiles/dev/client.yml
@@ -17,7 +17,7 @@ application_config:
   environment : "dev"
 
 registries :
-  "hangzhouzk":
+  "demoZk":
     protocol: "zookeeper"
     timeout	: "3s"
     address: "127.0.0.1:2181"
@@ -28,7 +28,7 @@ registries :
 references:
   "UserProvider":
     # 可以指定多个registry，使用逗号隔开;不指定默认向所有注册中心注册
-    registry: "hangzhouzk"
+    registry: "demoZk"
     protocol : "dubbo"
     interface : "com.ikurento.user.UserProvider"
     cluster: "failover"
