+++ examples/helloworld/dubbo/go-server/profiles/dev/server.yml
@@ -11,20 +11,15 @@ application_config:
   environment : "dev"
 
 registries :
-  "hangzhouzk":
-    # 对应java配置中address属性的zookeeper <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
+  "demoZk":
     protocol: "zookeeper"
     timeout	: "3s"
     address: "127.0.0.1:2181"
-    username: ""
-    password: ""
-
-
 
 services:
   "UserProvider":
     # 可以指定多个registry，使用逗号隔开;不指定默认向所有注册中心注册
-    registry: "hangzhouzk"
+    registry: "demoZk"
     protocol : "dubbo"
     # 相当于dubbo.xml中的interface
     interface : "com.ikurento.user.UserProvider"
@@ -36,11 +31,9 @@ services:
       retries: 1
       loadbalance: "random"
 
-
 protocols:
-  "dubbo1":
+  "dubbo":
     name: "dubbo"
-  #  ip : "127.0.0.1"
     port: 20000
